YBC

This submodule enables interaction with the "YBC" bolometric correction (BC) grid described in Chen et al. (2019). These BCs are commonly used in conjunction with the PARSEC stellar models and are used in the "CMD webform" maintained by Leo Girardi. The YBC grid presents a uniform processing of a number of different stellar atmosphere libraries (see section 3.1 of Chen et al. (2019)) that each cover different parameter spaces (e.g., surface gravity and effective temperature) as well as different types of stars (e.g., Wolf-Rayet stars, white dwarfs). This approach allows YBC to be robust across a broad range of stellar evolutionary states and makes it very useful for generating synthetic stellar photometry. However, many of these libraries are only used when particular stellar conditions are met, and not all libraries cover the same range of stellar population parameters (metallicity in particular).

Our implementation currently only supports a subset of the libraries in the full YBC set. In particular, we support the PHOENIX models which are used for cool stars, and are working to support the ATLAS9 models used for hotter stars.

Obtaining Data

Our naming convention for photometric filter systems follows the directory structure of the YBC git repository – see supported systems below. This repository can be accessed at the URL below.

https://gitlab.com/cycyustc/ybc_tables

The data are obtained by creating a local sparse clone of the YBC git repository with Scratch.jl and only pulling the data for photometric systems that you request. Presently we only support the standard BCs hosted under the "YBC" subdirectory. These data will be removed automatically if you uninstall the package. In the event you wish to uninstall the data for a particular photometric filter system, you can use BolometricCorrections.YBC.remove_table, though this should not typically be necessary as the data for each system is on average ~20 MB.

BolometricCorrections.YBC.remove_table — Function

remove_table(f::AbstractString, prefix::AbstractString = "YBC")

Remove table joinpath(prefix, f) from the Git sparse-checkout list for YBC – this will uninstall the related data files.

Supported Systems

Supported filter systems available at the time these docs were built are listed below.

2mass               decam                       hipparcos                     lsst_gaiaDR2maiz          suprimecam               wfc3_medium
2mass_spitzer       decam_vircam                hsc                           lsst_wfirst_notional      swift_uvot               wfc3_uvisCaHK
2mass_spitzer_wise  deltaa                      int_wfc                       lsst_wfirst_proposed2017  tycho2                   wfc3_verywide
CSST                denis                       iphas                         megacam                   ubvrijhk                 wfc3_wide
LUVIT               dmc14                       jpas                          megacam_post2014          ubvrijhklm               wfc3_wideverywide
Roman2021           dmc15                       jwst_fnl                      megacam_wircam            ukidss                   wfi
Roman2024           dualclump                   jwst_miri_wide                nicmosab                  uvit                     wfi2
TESS_2mass          duets                       jwst_narrow                   nicmosvega                vertecs                  wfirst
TESS_2mass_kepler   eis                         jwst_nircam_wide              noao_ctio_mosaic2         vilnius                  wfirst_J_H_1
UVbright            euclid_decam                jwst_nircam_widemedium        ogle                      vircam                   wfirst_J_H_2
WFIRST_plusHST      euclid_nisp                 jwst_nircam_widemedium_nov22  ogle_2mass_spitzer        visir                    wfirst_J_H_3
acs_2mass_spitzer   gaia                        jwst_niriss_nov22             panstarrs1                vista                    wfirst_J_H_7_1
acs_hrc             gaiaDR2                     jwst_nirspec                  phat_agb                  vista_ukidss             wfirst_J_H_7_2
acs_wfc             gaiaDR2_tycho2_2mass        jwst_wide                     phat_small                vphas                    wfirst_notional
acs_wfc_202101      gaiaDR2maiz                 keck_nirc2                    phat_spitzer              vst_omegacam             wfirst_plusHST
acs_wfc_pos04jul06  gaiaDR2weiler               kepler                        sage                      vst_vista                wfirst_proposed2017
acs_wfc_pre04jul06  gaiaDR2weiler_tycho2_2mass  kepler_2mass                  skymapper                 washington_ddo51         wfirst_shift_2
akari               gaiaDR3sim                  lbt_lbc                       sloan                     wasp                     wfirst_shift_3
batc                gaiaEDR3                    lsst                          sloan_2mass               wfc3_202101_IR_narrow    wfpc2
bessell             gaia_tycho2_2mass           lsstDP0                       sloan_ukidss              wfc3_202101_UVIS_narrow  wircam
caspir              galex                       lsstR1.9                      spitzer                   wfc3_202101_medium       wise
ciber               galex_sloan                 lsstR1.9_gaiaEDR3_euclid      splus                     wfc3_202101_verywide     ztf
ciber2              galexgaia                   lsst_gaia                     stis                      wfc3_202101_wide
clue_galex          gulli                       lsst_gaiaDR2                  stroemgren                wfc3_UVIS_narrow

YBC References

This page cites the following references:

Chen, Y.; Girardi, L.; Fu, X.; Bressan, A.; Aringer, B.; Dal Tio, P.; Pastorelli, G.; Marigo, P.; Costa, G. and Zhang, X. (2019). YBC: a stellar bolometric corrections database with variable extinction coefficients. Application to PARSEC isochrones. A & A 632, A105, arXiv:1910.09037 [astro-ph.SR].